refactor(LottoPrizeAnswer..): 당첨정답 비교 기능 리팩토링
- LottoWinningNumbers가 담당하던 정답과의 비교를 LottoPrizeAnswer로 역할 변경
- LottoPrizeAnswer는 LottoWinningNumbers와 BonusNumber와의 비교를 함께 하여 결과 제출
