+++ src/main/java/lottoservice/matcher/LottoPrizeAnswer.java
@@ -28,26 +28,11 @@ public class LottoPrizeAnswer {
 	public LottoMatchResult matchTickets(LottoTickets lottoTickets) {
 		LottoMatchResult lottoMatchResult = new LottoMatchResult();
 		for (LottoTicket lottoTicket : lottoTickets.getLottoTickets()) {
+			lottoMatchResult.addMatchCount(matchTicket(lottoTicket));
 			// setMatchResult(lottoMatchResult, lottoTicket);
 		}
 		return lottoMatchResult;
 	}
-
-	// /* 인자로 넘어온 티켓들을 정답과 비교하여 결과를 리턴 */
-	// public LottoMatchResult matchWinningAndTickets(LottoTickets lottoTickets) {
-	//
-	// 	for (LottoTicket lottoTicket : lottoTickets.getLottoTickets()) {
-	// 		setMatchResult(lottoMatchResult, lottoTicket);
-	// 	}
-	// 	return lottoMatchResult;
-	// }
-	//
-	// private void setMatchResult(LottoMatchResult lottoMatchResult, LottoTicket lottoTicket) {
-	// 	int matchCount = matchCountWinningAndTicket(lottoTicket);
-	// 	if (hasMatch(matchCount)) {
-	// 		lottoMatchResult.addMatchCount(LottoMatchRank.valueOf(matchCount,false));
-	// 	}
-	// }
 
 	public LottoMatchRank matchTicket(LottoTicket lottoTicket) {
 		boolean matchBonus = bonusNumber.matchTicket(lottoTicket);
